requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
djangorestframework==3.11.1
idna==2.10
numpy==1.19.2
opencv-python==4.4.0.44
pytz==2020.1
pyzbar==0.1.8
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
Wand==0.6.3
